For testing.
This creates a list of products that each has a supplier connected.
The products has then one or more offers connected.
If you startup the application and activate the product controller you
will get a list of products and if you click the red marked volume to the right
then the connected offers will be listed below.
But if you use the select to select a supplier then you can not update the offer list any more until you refresh the page.